// frontend_zh/src/components/thinkflow/types.ts
export type Notebook = {
  id: string;
  title?: string;
  name?: string;
};